EXECUTIVE_AI_ASSISTANT/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# AI and ML
openai==1.3.0
anthropic==0.7.0
langchain==0.0.335
transformers==4.35.0

# Voice and Speech
speechrecognition==3.10.0
pyttsx3==2.90
# pyaudio==0.2.13  # Commented out - requires system dependencies
gtts==2.4.0

# Web and API
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0

# CLI and UI
rich==13.7.0

# Data processing
pandas==2.1.3
numpy==1.26.0
scipy==1.11.4

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0